+++ ftc-robotcontroller/src/main/java/com/lasarobotics/ftcrobotcontroller/opmodes/FtcOpModeRegister.java
@@ -48,7 +48,6 @@ public class FtcOpModeRegister implements OpModeRegister {
     public void register(OpModeManager manager) {
         //Custom op modes
         manager.register("Null", NullOp.class);
-        manager.register("Vision Test 1", VisionTest1.class);
-        manager.register("Vision Test 2", VisionTest2.class);
+        manager.register("Vision Test", VisionTest.class);
     }
 }
