--- ftc-robotcontroller/src/main/java/com/lasarobotics/ftcrobotcontroller/opmodes/VisionTest2.java
@@ -1,120 +1,0 @@
-/* Copyright (c) 2014, 2015 Qualcomm Technologies Inc
-
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted (subject to the limitations in the disclaimer below) provided that
-the following conditions are met:
-
-Redistributions of source code must retain the above copyright notice, this list
-of conditions and the following disclaimer.
-
-Redistributions in binary form must reproduce the above copyright notice, this
-list of conditions and the following disclaimer in the documentation and/or
-other materials provided with the distribution.
-
-Neither the name of Qualcomm Technologies Inc nor the names of its contributors
-may be used to endorse or promote products derived from this software without
-specific prior written permission.
-
-NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
-LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
-THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
-
-package com.lasarobotics.ftcrobotcontroller.opmodes;
-
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-
-import org.lasarobotics.vision.detection.ColorBlobDetector;
-import org.lasarobotics.vision.detection.objects.Contour;
-import org.lasarobotics.vision.ftc.resq.Beacon;
-import org.lasarobotics.vision.image.Transform;
-import org.lasarobotics.vision.ui.VisionEnabledActivity;
-import org.lasarobotics.vision.ui.VisionOpMode;
-import org.lasarobotics.vision.util.FPS;
-import org.lasarobotics.vision.util.color.ColorHSV;
-import org.opencv.android.CameraBridgeViewBase;
-import org.opencv.core.CvType;
-import org.opencv.core.Mat;
-
-import java.util.List;
-
-/**
- * TeleOp Mode
- * <p/>
- * Enables control of the robot via the gamepad
- */
-public class VisionTest2 extends VisionOpMode {
-
-    Beacon.BeaconColorAnalysis colorAnalysis = new Beacon.BeaconColorAnalysis();
-    private ColorBlobDetector detectorRed;
-    private ColorBlobDetector detectorBlue;
-    private static final ColorHSV lowerBoundRed = new ColorHSV((int) (305 / 360.0 * 255.0), (int) (0.200 * 255.0), (int) (0.300 * 255.0));
-    private static final ColorHSV upperBoundRed = new ColorHSV((int) ((360.0 + 5.0) / 360.0 * 255.0), 255, 255);
-    private static final ColorHSV lowerBoundBlue = new ColorHSV((int) (170.0 / 360.0 * 255.0), (int) (0.200 * 255.0), (int) (0.750 * 255.0));
-    private static final ColorHSV upperBoundBlue = new ColorHSV((int) (227.0 / 360.0 * 255.0), 255, 255);
-
-    @Override
-    public void init(int width, int height) {
-        //Initialize all detectors here
-        detectorRed = new ColorBlobDetector(lowerBoundRed, upperBoundRed);
-        detectorBlue = new ColorBlobDetector(lowerBoundBlue, upperBoundBlue);
-    }
-
-    /*
-     * This method will be called repeatedly in a loop
-     * @see com.qualcomm.robotcore.eventloop.opmode.OpMode#loop()
-     */
-    @Override
-    public void loop() {
-        telemetry.addData("Vision Color", colorAnalysis.toString());
-    }
-
-    @Override
-    public Mat frame(Mat rgba, Mat gray) {
-        //DEBUG for the Nexus
-        //Transform.flip(mRgba, Transform.FlipType.FLIP_BOTH);
-        //Transform.flip(mGray, Transform.FlipType.FLIP_BOTH);
-
-        //Transform.shrink(mRgba, new Size(480, 480), true);
-        //Transform.shrink(mGray, new Size(480, 480), true);
-
-        //DEBUG for the Moto G
-        Transform.rotate(gray, -90);
-        Transform.rotate(rgba, -90);
-
-        try {
-            //Process the frame for the color blobs
-            detectorRed.process(rgba);
-            detectorBlue.process(rgba);
-
-            //Get the list of contours
-            List<Contour> contoursRed = detectorRed.getContours();
-            List<Contour> contoursBlue = detectorBlue.getContours();
-
-            //Get color analysis
-            Beacon beacon = new Beacon(rgba.size());
-            colorAnalysis = beacon.analyzeColor(contoursRed, contoursBlue, rgba, gray);
-
-        } catch (Exception e) {
-            telemetry.addData("Vision Status", "Analysis Error");
-            e.printStackTrace();
-        }
-
-        telemetry.addData("Vision FPS", fps.getFPSString());
-        return rgba;
-    }
-
-    @Override
-    public void stop() {
-
-    }
-}
